pipeline/pipeline_trigger_function/requirements.txt
google-cloud-aiplatform==1.22.0
pyyaml
